fix import path for wav helper
