+++ src/main.ts
@@ -21,6 +21,7 @@ import { join } from 'path';
 import { readFileSync } from 'fs';
 import http from 'http';
 import crypto from 'crypto';
+import { createWavBuffer } from './wav.js';
 
 const { deviceManager } = sdk;
 
@@ -247,7 +248,7 @@ class BirdNETDevice extends ScryptedDeviceBase implements Settings {
     async analyzeAudioChunk(samples: Float32Array) {
         try {
             // Create WAV buffer from samples
-            const wavBuffer = this.createWavBuffer(samples);
+            const wavBuffer = createWavBuffer(samples);
             // Write temporary WAV file
             const os = require('os');
             const fs = require('fs');
@@ -288,38 +289,6 @@ class BirdNETDevice extends ScryptedDeviceBase implements Settings {
         }
     }
 
-    private createWavBuffer(samples: Float32Array): Buffer {
-        const numChannels = 1;
-        const sampleRate = 48000;
-        const bitsPerSample = 16;
-        const byteRate = sampleRate * numChannels * bitsPerSample / 8;
-        const blockAlign = numChannels * bitsPerSample / 8;
-        const dataLength = samples.length * 2; // 2 bytes per sample
-        const buffer = Buffer.alloc(44 + dataLength);
-        // RIFF header
-        buffer.write('RIFF', 0);
-        buffer.writeUInt32LE(36 + dataLength, 4);
-        buffer.write('WAVE', 8);
-        // fmt subchunk
-        buffer.write('fmt ', 12);
-        buffer.writeUInt32LE(16, 16);
-        buffer.writeUInt16LE(1, 20); // PCM
-        buffer.writeUInt16LE(numChannels, 22);
-        buffer.writeUInt32LE(sampleRate, 24);
-        buffer.writeUInt32LE(byteRate, 28);
-        buffer.writeUInt16LE(blockAlign, 32);
-        buffer.writeUInt16LE(bitsPerSample, 34);
-        // data subchunk
-        buffer.write('data', 36);
-        buffer.writeUInt32LE(dataLength, 40);
-        
-        for (let i = 0; i < samples.length; i++) {
-            let s = Math.max(-1, Math.min(1, samples[i]));
-            let intSample = Math.round(s * 32767);
-            buffer.writeInt16LE(intSample, 44 + i * 2);
-        }
-        return buffer;
-    }
 
     // Implement Web interface
     async getResource(requestBody: string): Promise<string> {
